additional functionality for testing settings and code cleanup
